[charts] Fix area appearing
