+++ packages/x-charts/src/LineChart/AnimatedArea.tsx
@@ -4,10 +4,9 @@ import PropTypes from 'prop-types';
 import { styled } from '@mui/material/styles';
 import { animated, useTransition } from '@react-spring/web';
 import { color as d3Color } from '@mui/x-charts-vendor/d3-color';
-import { cleanId } from '../internals/cleanId';
 import type { AreaElementOwnerState } from './AreaElement';
-import { useChartId, useDrawingArea } from '../hooks';
 import { useStringInterpolator } from '../internals/useStringInterpolator';
+import { AppearingMask } from './AppearingMask';
 
 export const AreaElementPath = styled(animated.path, {
   name: 'MuiAreaElement',
@@ -45,19 +44,8 @@ export interface AnimatedAreaProps extends React.ComponentPropsWithoutRef<'path'
  */
 function AnimatedArea(props: AnimatedAreaProps) {
   const { d, skipAnimation, ownerState, ...other } = props;
-  const { left, top, right, bottom, width, height } = useDrawingArea();
-  const chartId = useChartId();
 
   const stringInterpolator = useStringInterpolator(d);
-
-  const transitionAppear = useTransition([1], {
-    from: { animatedWidth: left },
-    to: { animatedWidth: width + left + right },
-    enter: { animatedWidth: width + left + right },
-    leave: { animatedWidth: left },
-    reset: false,
-    immediate: skipAnimation,
-  });
 
   const transitionChange = useTransition([stringInterpolator], {
     from: { value: 0 },
@@ -67,20 +55,12 @@ function AnimatedArea(props: AnimatedAreaProps) {
     immediate: skipAnimation,
   });
 
-  const clipId = cleanId(`${chartId}-${ownerState.id}-area-clip`);
   return (
-    <React.Fragment>
-      <clipPath id={clipId}>
-        {transitionAppear((style) => (
-          <animated.rect x={0} y={0} width={style.animatedWidth} height={top + height + bottom} />
-        ))}
-      </clipPath>
-      <g clipPath={`url(#${clipId})`}>
-        {transitionChange((style, interpolator) => (
-          <AreaElementPath {...other} ownerState={ownerState} d={style.value.to(interpolator)} />
-        ))}
-      </g>
-    </React.Fragment>
+    <AppearingMask skipAnimation={skipAnimation} id={`${ownerState.id}-area-clip`}>
+      {transitionChange((style, interpolator) => (
+        <AreaElementPath {...other} ownerState={ownerState} d={style.value.to(interpolator)} />
+      ))}
+    </AppearingMask>
   );
 }
