+++ packages/x-charts/src/LineChart/AnimatedLine.tsx
@@ -4,11 +4,9 @@ import PropTypes from 'prop-types';
 import { animated, useTransition } from '@react-spring/web';
 import { color as d3Color } from '@mui/x-charts-vendor/d3-color';
 import { styled } from '@mui/material/styles';
-import { cleanId } from '../internals/cleanId';
 import type { LineElementOwnerState } from './LineElement';
-import { useChartId } from '../hooks/useChartId';
-import { useDrawingArea } from '../hooks/useDrawingArea';
 import { useStringInterpolator } from '../internals/useStringInterpolator';
+import { AppearingMask } from './AppearingMask';
 
 export const LineElementPath = styled(animated.path, {
   name: 'MuiLineElement',
@@ -48,21 +46,8 @@ export interface AnimatedLineProps extends React.ComponentPropsWithoutRef<'path'
  */
 function AnimatedLine(props: AnimatedLineProps) {
   const { d, skipAnimation, ownerState, ...other } = props;
-  const drawingArea = useDrawingArea();
-  const chartId = useChartId();
 
   const stringInterpolator = useStringInterpolator(d);
-
-  const transitionAppear = useTransition<typeof drawingArea, { animatedWidth: number }>(
-    [drawingArea],
-    {
-      from: (v) => ({ animatedWidth: v.left }),
-      enter: (v) => ({ animatedWidth: v.width + v.left + v.right }),
-      leave: (v) => ({ animatedWidth: v.width + v.left + v.right }),
-      reset: false,
-      immediate: skipAnimation,
-    },
-  );
 
   const transitionChange = useTransition([stringInterpolator], {
     from: { value: 0 },
@@ -72,25 +57,12 @@ function AnimatedLine(props: AnimatedLineProps) {
     immediate: skipAnimation,
   });
 
-  const clipId = cleanId(`${chartId}-${ownerState.id}-line-clip`);
   return (
-    <React.Fragment>
-      <clipPath id={clipId}>
-        {transitionAppear((style) => (
-          <animated.rect
-            x={0}
-            y={0}
-            width={style.animatedWidth}
-            height={drawingArea.top + drawingArea.height + drawingArea.bottom}
-          />
-        ))}
-      </clipPath>
-      <g clipPath={`url(#${clipId})`}>
-        {transitionChange((style, interpolator) => (
-          <LineElementPath {...other} ownerState={ownerState} d={style.value.to(interpolator)} />
-        ))}
-      </g>
-    </React.Fragment>
+    <AppearingMask skipAnimation={skipAnimation} id={`${ownerState.id}-line-clip`}>
+      {transitionChange((style, interpolator) => (
+        <LineElementPath {...other} ownerState={ownerState} d={style.value.to(interpolator)} />
+      ))}
+    </AppearingMask>
   );
 }
